requirements.txt
jinja2>=3.1.2
markdown>=3.5
pyyaml>=6.0
